% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grassland}
\alias{grassland}
\title{Mediterranean grassland: individual-level traits}
\format{
A data frame with 2,444 rows and 5 columns:
\describe{
\item{plot}{Plot identifier (integer, 1-40).}
\item{species}{Species name.}
\item{individual}{Individual number within plot and species (1-10).}
\item{height}{Vegetative height (cm).}
\item{sla}{Specific leaf area (mm2 mg-1).}
}
}
\source{
Carmona, C.P., Rota, C., Azcarate, F.M. & Peco, B. (2015)
More for less: sampling strategies of plant functional traits across
local environmental gradients. \emph{Functional Ecology}, 29, 579-588.
\doi{10.1111/1365-2435.12366}. See also \link{grassland_abund} for the
corresponding relative covers.
}
\usage{
grassland
}
\description{
Individual-level trait measurements from a Mediterranean annual
grassland (Valdeloshielos, central Spain). In each of 40 plots, up to
ten individuals of the most abundant species were measured. Only
individuals with both traits measured are included.
}
\keyword{datasets}
