% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grassland_abund}
\alias{grassland_abund}
\title{Mediterranean grassland: relative covers}
\format{
A numeric matrix with 40 rows (plots) and 51 columns (species).
}
\source{
Carmona, C.P., Rota, C., Azcarate, F.M. & Peco, B. (2015)
\emph{Functional Ecology}, 29, 579-588. \doi{10.1111/1365-2435.12366}.
}
\usage{
grassland_abund
}
\description{
Relative cover of each species in each of the 40 plots of the
Valdeloshielos grassland (rows sum to 1). Companion to \link{grassland}.
}
\keyword{datasets}
