% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gspff}
\alias{gspff}
\title{Global spectrum of plant form and function: complete trait data}
\format{
A data frame with 2,630 rows and 6 columns: \code{la} (leaf area),
\code{ln} (leaf nitrogen content), \code{ph} (plant height), \code{sla} (specific
leaf area), \code{ssd} (specific stem density), \code{sm} (seed mass).
}
\source{
\doi{10.6084/m9.figshare.13140146}; TRY database
(https://www.try-db.org). Carmona et al. (2021) \emph{Nature}, 597,
683-687.
}
\usage{
gspff
}
\description{
Six aboveground traits for 2,630 plant species with complete trait
information, from the global spectrum of plant form and function
(Diaz et al. 2016), as compiled from the TRY database and distributed
with the \pkg{funspace} package. All traits are log10-transformed and
scaled.
}
\keyword{datasets}
