% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gspff_missing}
\alias{gspff_missing}
\title{Global spectrum of plant form and function: incomplete trait data}
\format{
A data frame with 10,746 rows and 6 columns; see \link{gspff}.
}
\source{
\doi{10.6084/m9.figshare.13140146}.
}
\usage{
gspff_missing
}
\description{
The same six aboveground traits for 10,746 species with incomplete
trait information (species with at least three traits measured);
log10-transformed and scaled. Useful together with \link{gspff_phylo} to
demonstrate imputation with \code{\link[=fs_impute]{fs_impute()}}.
}
\keyword{datasets}
