% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gspff_phylo}
\alias{gspff_phylo}
\title{Phylogeny for the incomplete global spectrum species}
\format{
An object of class \code{phylo} with 10,746 tips.
}
\source{
Distributed with the \pkg{funspace} package.
}
\usage{
gspff_phylo
}
\description{
Phylogenetic tree (class \code{phylo}, \pkg{ape}) for the 10,746 species in
\link{gspff_missing}.
}
\keyword{datasets}
