% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gspff_tax}
\alias{gspff_tax}
\title{Taxonomy for the complete global spectrum species}
\format{
A data frame with 2,630 rows and 3 columns: \code{genus}, \code{family},
\code{order}.
}
\source{
\doi{10.6084/m9.figshare.13140146}.
}
\usage{
gspff_tax
}
\description{
Genus, family and order for the 2,630 species in \link{gspff}.
}
\keyword{datasets}
