% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fspace.R
\name{is_fspace}
\alias{is_fspace}
\title{Test whether an object is an fspace}
\usage{
is_fspace(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} inherits from class \code{fspace}.
}
\description{
Test whether an object is an fspace
}
\examples{
data(gspff)
is_fspace(fs_space(gspff[1:50, ], method = "pca"))
is_fspace(gspff)
}
